requirements.txt
bash==0.6
dj-rest-auth==4.0.1
Django==4.2.1
django-allauth==0.54.0
djangorestframework==3.14.0
environs==8.0.0
Pillow==9.5.0
psycopg2==2.9.5
requests==2.28.2
requests-oauthlib==1.3.1


